@@ -1,205 +1,216 @@
-﻿using EasyCLib.NET.Sdk;
+using EasyCLib.NET.Sdk;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Cors;
-using System.Data;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using Microsoft.AspNetCore.Authorization;
+using DataView = System.Data.DataView;
+using DataColumn = System.Data.DataColumn;
+using DataRowView = System.Data.DataRowView;
 
 namespace OHotel.NETCoreMVC.Controllers.API
 {
-    //[Route("api/[controller]")]
-    // 以下是命名路徑位置: (例如) Values/GetClass 而完整的話長得像是這樣: https:// localhost:44367/api/EHotelFood/GetEHotel_Food 的樣子
     [Route("api/[controller]/[action]")]
     [ApiController]
     [Authorize]
     public class LayoutController : ControllerBase
     {
-        private IDbFunction _IDbFunction;
-        public IConfiguration _Configuration { get; set; }
-        private String Sql { get; set; } = "";
+        private readonly IDbFunction _IDbFunction;
+        private readonly IConfiguration _Configuration;
+
         public LayoutController(IDbFunction dbFunction, IConfiguration configuration)
         {
             _IDbFunction = dbFunction;
             _Configuration = configuration;
         }
 
-        /// <summary>
-        /// ※查詢所有資料
-        /// 230503 Added by KEVINN （Origin from 【EHotelFoodController】）
-        /// </summary>
-        [EnableCors("CorsPolicy")] // 加入跨網權限設定
+        private bool IsSqlite => string.Equals(_Configuration["DatabaseProvider"] ?? "", "Sqlite", StringComparison.OrdinalIgnoreCase);
+
+        /// <summary>※查詢所有 ManageClass</summary>
+        [EnableCors("CorsPolicy")]
         [HttpGet]
-        public async Task<IActionResult> GetManageClass()
-        { // 使用SqlReader
-            string connectionString = _Configuration.GetConnectionString("SQLCD_Read_OHotel");
-            using (SqlConnection connection = new SqlConnection(connectionString))
+        public IActionResult GetManageClass()
+        {
+            var connStr = _Configuration.GetConnectionString("SQLCD_Read_OHotel");
+            if (IsSqlite)
             {
-                string sql = "SELECT * FROM ManageClass WHERE State = 0 order by ManageClass.[Order] ASC";
-                using (SqlCommand command = new SqlCommand(sql, connection))
+                _IDbFunction.DbConnect(connStr ?? "");
+                if (!_IDbFunction.SelectDbDataView("SELECT * FROM ManageClass WHERE State = 0 ORDER BY MCNo ASC", "ManageClass"))
                 {
-                    await connection.OpenAsync();
-                    using (SqlDataReader reader = await command.ExecuteReaderAsync())
-                    {
-                        List<IDictionary<string, object>> results = new List<IDictionary<string, object>>();
-                        while (await reader.ReadAsync())
-                        {
-                            var result = Enumerable.Range(0, reader.FieldCount).ToDictionary(reader.GetName, reader.GetValue);
-                            results.Add(result);
-                        }
-                        return Ok(results);
-                    }
+                    _IDbFunction.DbClose();
+                    return Ok(new List<IDictionary<string, object>>());
                 }
+                var results = DataViewToList(_IDbFunction.SqlDataView);
+                _IDbFunction.DbClose();
+                return Ok(results);
             }
+            return RunSqlServer(async (conn) =>
+            {
+                using var cmd = new SqlCommand("SELECT * FROM ManageClass WHERE State = 0 ORDER BY ManageClass.[Order] ASC", conn);
+                using var reader = await cmd.ExecuteReaderAsync();
+                return Ok(ReaderToList(reader));
+            });
         }
 
-        /// <summary>
-        /// ※獲取登入的會員資訊
-        /// </summary>
+        /// <summary>※獲取登入的會員名稱</summary>
         [EnableCors("CorsPolicy")]
         [HttpGet("{STNo}")]
         public IActionResult GetSTNameBySTNo(int STNo)
         {
-            try
+            if (IsSqlite)
             {
-                string connectionString = _Configuration.GetConnectionString("SQLCD_Read_OHotel");
-                using (SqlConnection connection = new SqlConnection(connectionString))
-                {
-                    string query = "SELECT STName FROM Staff WHERE STNo = @STNo";
-                    using (SqlCommand command = new SqlCommand(query, connection))
-                    {
-                        command.Parameters.AddWithValue("@STNo", STNo);
-
-                        connection.Open();
-                        using (SqlDataReader reader = command.ExecuteReader())
-                        {
-                            if (reader.Read())
-                            {
-                                string STName = reader.GetString(reader.GetOrdinal("STName"));
-                                return Ok(STName);
-                            }
-                            else
-                            {
-                                return NotFound("Unknown");
-                            }
-                        }
-                    }
-                }
+                var connStr = _Configuration.GetConnectionString("SQLCD_Read_OHotel");
+                _IDbFunction.DbConnect(connStr ?? "");
+                var ok = _IDbFunction.SelectDbDataViewWithParams("SELECT STName FROM Staff WHERE STNo = @STNo", "Staff", new Dictionary<string, object> { ["@STNo"] = STNo });
+                _IDbFunction.DbClose();
+                if (ok && _IDbFunction.SqlDataView.Count > 0)
+                    return Ok(_IDbFunction.SqlDataView[0]["STName"]?.ToString() ?? "Unknown");
+                return NotFound("Unknown");
             }
-            catch (Exception ex)
+            return RunSqlServerSync((conn) =>
             {
-                Console.WriteLine("An error occurred while retrieving STName: " + ex.Message);
-                return StatusCode(500, "An error occurred while retrieving STName");
-            }
+                using var cmd = new SqlCommand("SELECT STName FROM Staff WHERE STNo = @STNo", conn);
+                cmd.Parameters.AddWithValue("@STNo", STNo);
+                using var reader = cmd.ExecuteReader();
+                return reader.Read() ? Ok(reader.GetString(0)) : NotFound("Unknown");
+            });
         }
 
-        /// <summary>
-        /// ※取得SIDEBAR標題 這個後續要獨立出來
-        /// 230510 Created by KEVINN
-        /// </summary>
+        /// <summary>※取得 SIDEBAR 選單（依權限）</summary>
         [EnableCors("CorsPolicy")]
         [HttpGet("{STNo}")]
-        public async Task<IActionResult> GetManageClassAndItemsBySTNo(int STNo)
+        public IActionResult GetManageClassAndItemsBySTNo(int STNo)
         {
-            string connectionString = _Configuration.GetConnectionString("SQLCD_Read_OHotel");
-            using (SqlConnection connection = new SqlConnection(connectionString))
+            var connStr = _Configuration.GetConnectionString("SQLCD_Read_OHotel");
+            if (IsSqlite)
             {
+                _IDbFunction.DbConnect(connStr ?? "");
+                var allPowerOk = _IDbFunction.SelectDbDataViewWithParams("SELECT AllPower FROM Staff WHERE STNo = @STNo", "Staff", new Dictionary<string, object> { ["@STNo"] = STNo });
+                var allPower = (allPowerOk && _IDbFunction.SqlDataView.Count > 0) ? Convert.ToInt32(_IDbFunction.SqlDataView[0]["AllPower"]) : 0;
                 string sql;
-                bool allPowerEqualsZero;
-
-                // Check the value of AllPower in the Staff table
-                using (SqlCommand checkCommand = new SqlCommand("SELECT AllPower FROM Staff WHERE STNo = @STNo", connection))
+                bool ok;
+                if (allPower == 0)
                 {
-                    checkCommand.Parameters.AddWithValue("@STNo", STNo);
-                    await connection.OpenAsync();
-                    var allPowerResult = await checkCommand.ExecuteScalarAsync();
-                    allPowerEqualsZero = Convert.ToInt32(allPowerResult) == 0;
-                }
-
-                if (allPowerEqualsZero)
-                {
-                    sql = @"
-                SELECT MC.MCNo, MC.MCName, MC.MCIcon, MC.MCXtrol, MI.MIAction, MI.ItemName
-                FROM ManageClass MC
-                INNER JOIN ManageItem MI ON MC.MCNo = MI.MCNo
-                WHERE MC.State = 0
-                    AND MC.MCNo IN (
-                        SELECT MCNo
-                        FROM ManageItem
-                        WHERE ManageItem.MINo IN (
-                            SELECT MINo
-                            FROM StaffPower
-                            WHERE STNo = @STNo AND PV = 1
-                        ) AND ManageItem.State = 0
-                    )
-                    AND MI.State = 0
-                    AND MI.MINo IN (
-                        SELECT MINo
-                        FROM StaffPower
-                        WHERE STNo = @STNo AND PV = 1
-                    )
-                ORDER BY MC.[Order] ASC";
+                    sql = @"SELECT MC.MCNo, MC.MCName, MC.MCIcon, MC.MCXtrol, MI.MIAction, MI.ItemName
+FROM ManageClass MC INNER JOIN ManageItem MI ON MC.MCNo = MI.MCNo
+WHERE MC.State = 0 AND MI.State = 0 AND MI.MINo IN (SELECT MINo FROM StaffPower WHERE STNo = @STNo AND PV = 1)
+ORDER BY MC.MCNo ASC, MI.MINo ASC";
+                    ok = _IDbFunction.SelectDbDataViewWithParams(sql, "Menu", new Dictionary<string, object> { ["@STNo"] = STNo });
                 }
                 else
                 {
-                    sql = @"
-                SELECT MC.MCNo, MC.MCName, MC.MCIcon, MC.MCXtrol, MI.MIAction, MI.ItemName
-                FROM ManageClass MC
-                INNER JOIN ManageItem MI ON MC.MCNo = MI.MCNo
-                WHERE MC.State = 0
-                    AND MI.State = 0
-                ORDER BY MC.[Order] ASC";
+                    sql = @"SELECT MC.MCNo, MC.MCName, MC.MCIcon, MC.MCXtrol, MI.MIAction, MI.ItemName
+FROM ManageClass MC INNER JOIN ManageItem MI ON MC.MCNo = MI.MCNo
+WHERE MC.State = 0 AND MI.State = 0
+ORDER BY MC.MCNo ASC, MI.MINo ASC";
+                    ok = _IDbFunction.SelectDbDataView(sql, "Menu");
                 }
-
-                using (SqlCommand command = new SqlCommand(sql, connection))
+                var results = new Dictionary<string, List<IDictionary<string, object>>>();
+                if (ok && _IDbFunction.SqlDataView.Count > 0)
                 {
-                    command.Parameters.AddWithValue("@STNo", STNo);
-                    using (SqlDataReader reader = await command.ExecuteReaderAsync())
+                    foreach (DataRowView row in _IDbFunction.SqlDataView)
                     {
-                        Dictionary<string, List<IDictionary<string, object>>> results = new Dictionary<string, List<IDictionary<string, object>>>();
-                        while (await reader.ReadAsync())
-                        {
-                            string MCName = reader.GetString(reader.GetOrdinal("MCName"));
-                            if (!results.ContainsKey(MCName))
-                            {
-                                results[MCName] = new List<IDictionary<string, object>>();
-                            }
-                            var result = Enumerable.Range(0, reader.FieldCount).ToDictionary(reader.GetName, reader.GetValue);
-                            results[MCName].Add(result);
-                        }
-                        return Ok(results);
+                        var mcName = row["MCName"]?.ToString() ?? "";
+                        if (!results.ContainsKey(mcName)) results[mcName] = new List<IDictionary<string, object>>();
+                        var dict = new Dictionary<string, object>();
+                        var tbl = _IDbFunction.SqlDataView.Table;
+                        if (tbl != null)
+                            foreach (DataColumn col in tbl.Columns)
+                            dict[col.ColumnName] = row[col.ColumnName] ?? DBNull.Value;
+                        results[mcName].Add(dict);
                     }
                 }
+                _IDbFunction.DbClose();
+                return Ok(results);
             }
+            return RunSqlServer(async (conn) =>
+            {
+                using var checkCmd = new SqlCommand("SELECT AllPower FROM Staff WHERE STNo = @STNo", conn);
+                checkCmd.Parameters.AddWithValue("@STNo", STNo);
+                var allPower = Convert.ToInt32(await checkCmd.ExecuteScalarAsync());
+                string sql = allPower == 0
+                    ? @"SELECT MC.MCNo, MC.MCName, MC.MCIcon, MC.MCXtrol, MI.MIAction, MI.ItemName
+FROM ManageClass MC INNER JOIN ManageItem MI ON MC.MCNo = MI.MCNo
+WHERE MC.State = 0 AND MC.MCNo IN (SELECT MCNo FROM ManageItem WHERE MINo IN (SELECT MINo FROM StaffPower WHERE STNo = @STNo AND PV = 1) AND State = 0)
+AND MI.State = 0 AND MI.MINo IN (SELECT MINo FROM StaffPower WHERE STNo = @STNo AND PV = 1)
+ORDER BY MC.[Order] ASC"
+                    : @"SELECT MC.MCNo, MC.MCName, MC.MCIcon, MC.MCXtrol, MI.MIAction, MI.ItemName
+FROM ManageClass MC INNER JOIN ManageItem MI ON MC.MCNo = MI.MCNo
+WHERE MC.State = 0 AND MI.State = 0
+ORDER BY MC.[Order] ASC";
+                using var cmd = new SqlCommand(sql, conn);
+                cmd.Parameters.AddWithValue("@STNo", STNo);
+                using var reader = await cmd.ExecuteReaderAsync();
+                var results = new Dictionary<string, List<IDictionary<string, object>>>();
+                while (await reader.ReadAsync())
+                {
+                    var mcName = reader.GetString(reader.GetOrdinal("MCName"));
+                    if (!results.ContainsKey(mcName)) results[mcName] = new List<IDictionary<string, object>>();
+                    results[mcName].Add(Enumerable.Range(0, reader.FieldCount).ToDictionary(reader.GetName, reader.GetValue));
+                }
+                return Ok(results);
+            });
         }
 
-        /// <summary>
-        /// ※透過編號查詢
-        /// 230508 Added by KEVINN （Origin from 【EHotelFoodController】）
-        /// </summary>
+        /// <summary>※透過編號查詢 ManageClass</summary>
         [EnableCors("CorsPolicy")]
         [HttpGet("MCNo/{MCNo}")]
-        public async Task<IActionResult> GetManageClass_ByMCNo(int MCNo)
+        public IActionResult GetManageClass_ByMCNo(int MCNo)
         {
-            string connectionString = _Configuration.GetConnectionString("SQLCD_Read_OHotel");
-            using (SqlConnection connection = new SqlConnection(connectionString))
+            if (IsSqlite)
             {
-                string sql = @"SELECT * FROM ManageClass WHERE MCNo = @MCNo";
-                using (SqlCommand command = new SqlCommand(sql, connection))
-                {
-                    command.Parameters.AddWithValue("@MCNo", MCNo);
-                    await connection.OpenAsync();
-                    using (SqlDataReader reader = await command.ExecuteReaderAsync())
-                    {
-                        List<IDictionary<string, object>> results = new List<IDictionary<string, object>>();
-                        while (await reader.ReadAsync())
-                        {
-                            var result = Enumerable.Range(0, reader.FieldCount).ToDictionary(reader.GetName, reader.GetValue);
-                            results.Add(result);
-                        }
-                        return Ok(results);
-                    }
-                }
+                var connStr = _Configuration.GetConnectionString("SQLCD_Read_OHotel");
+                _IDbFunction.DbConnect(connStr ?? "");
+                var ok = _IDbFunction.SelectDbDataViewWithParams("SELECT * FROM ManageClass WHERE MCNo = @MCNo", "MC", new Dictionary<string, object> { ["@MCNo"] = MCNo });
+                _IDbFunction.DbClose();
+                return Ok(ok && _IDbFunction.SqlDataView.Count > 0 ? DataViewToList(_IDbFunction.SqlDataView) : new List<IDictionary<string, object>>());
             }
+            return RunSqlServer(async (conn) =>
+            {
+                using var cmd = new SqlCommand("SELECT * FROM ManageClass WHERE MCNo = @MCNo", conn);
+                cmd.Parameters.AddWithValue("@MCNo", MCNo);
+                using var reader = await cmd.ExecuteReaderAsync();
+                var list = ReaderToList(reader);
+                return Ok(list);
+            });
+        }
+
+        private static List<IDictionary<string, object>> DataViewToList(DataView dv)
+        {
+            var list = new List<IDictionary<string, object>>();
+            var table = dv?.Table;
+            if (table == null) return list;
+            for (int i = 0; i < dv!.Count; i++)
+            {
+                var dict = new Dictionary<string, object>();
+                foreach (DataColumn col in table.Columns)
+                    dict[col.ColumnName] = dv[i][col.ColumnName] ?? DBNull.Value;
+                list.Add(dict);
+            }
+            return list;
+        }
+
+        private static List<IDictionary<string, object>> ReaderToList(SqlDataReader reader)
+        {
+            var list = new List<IDictionary<string, object>>();
+            while (reader.Read())
+                list.Add(Enumerable.Range(0, reader.FieldCount).ToDictionary(reader.GetName, i => reader.GetValue(i)));
+            return list;
+        }
+
+        private IActionResult RunSqlServer(Func<SqlConnection, Task<IActionResult>> fn)
+        {
+            var connStr = _Configuration.GetConnectionString("SQLCD_Read_OHotel");
+            using var conn = new SqlConnection(connStr);
+            conn.Open();
+            return fn(conn).GetAwaiter().GetResult();
+        }
+
+        private IActionResult RunSqlServerSync(Func<SqlConnection, IActionResult> fn)
+        {
+            var connStr = _Configuration.GetConnectionString("SQLCD_Read_OHotel");
+            using var conn = new SqlConnection(connStr);
+            conn.Open();
+            return fn(conn);
         }
 
         private readonly string storagePath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Happy");
@@ -210,30 +221,31 @@ namespace OHotel.NETCoreMVC.Controllers.API
             try
             {
                 if (file == null || file.Length == 0)
-                {
-                    return BadRequest("No file uploaded.");
-                }
+                    return BadRequest("未上傳檔案");
 
-                string fileName = Guid.NewGuid().ToString(); // 使用唯一标识符作为文件名
+                const long maxFileSize = 10 * 1024 * 1024; // 10 MB
+                if (file.Length > maxFileSize)
+                    return BadRequest("檔案大小不可超過 10 MB");
 
-                // 获取文件扩展名
-                string fileExtension = Path.GetExtension(file.FileName);
+                var allowedExtensions = new[] { ".jpg", ".jpeg", ".png", ".gif", ".webp", ".pdf", ".doc", ".docx" };
+                var ext = Path.GetExtension(file.FileName)?.ToLowerInvariant();
+                if (string.IsNullOrEmpty(ext) || !allowedExtensions.Contains(ext))
+                    return BadRequest($"僅允許上傳以下格式: {string.Join(", ", allowedExtensions)}");
 
-                // 拼接文件名和扩展名
-                fileName += fileExtension;
-
-                string filePath = Path.Combine(storagePath, fileName);
+                var fileName = Guid.NewGuid().ToString() + ext;
+                var filePath = Path.Combine(storagePath, fileName);
+                Directory.CreateDirectory(storagePath);
 
                 using (var stream = new FileStream(filePath, FileMode.Create))
                 {
                     await file.CopyToAsync(stream);
                 }
 
-                return Ok(fileName); // 返回文件名或文件路径作为响应
+                return Ok(fileName);
             }
             catch (Exception ex)
             {
-                // 处理异常
+                // 處理例外
                 return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
             }
         }
