@@ -1,5 +1,5 @@
-﻿<header id="header" class="clearfix shadow">
-    <h1 class="logo"> <a asp-controller="Home" asp-action="Index"><img src="~/img/fp/logo.png" width="208" height="140" alt="飯店LOGO"></a></h1>
+<header id="header" class="clearfix shadow">
+    <h1 class="logo"> <a asp-area="FP-client" asp-controller="Home" asp-action="Index"><img src="~/img/fp/logo.png" width="208" height="140" alt="飯店LOGO"></a></h1>
     <!-- 選單 -->
     <div style="cursor:pointer" onclick="openNav()">
         <i class="fa fa-bars" aria-hidden="true"></i>
@@ -8,12 +8,12 @@
         <a href="javascript:void(0)" class="closebtn" onclick="closeNav()">&times;</a>
         <nav class="nav">
             <ul>
-                <li><a asp-controller="About" asp-action="Index" title="關於我們">關於我們</a></li>
-                <li><a asp-controller="Rooms" asp-action="Index" title="客房">客房</a></li>
-                <li><a asp-controller="Facility" asp-action="Index" title="設施">設施</a></li>
-                <li><a asp-controller="sights" asp-action="Index" title="週邊景點">週邊景點</a></li>
-                <li><a asp-controller="traffic" asp-action="Index" title="交通位置">交通位置</a></li>
-                <li><a asp-controller="news" asp-action="Index" title="最新消息">最新消息</a></li>
+                <li><a asp-area="FP-client" asp-controller="About" asp-action="Index" title="關於我們">關於我們</a></li>
+                <li><a asp-area="FP-client" asp-controller="Rooms" asp-action="Index" title="客房">客房</a></li>
+                <li><a asp-area="FP-client" asp-controller="Facility" asp-action="Index" title="設施">設施</a></li>
+                <li><a asp-area="FP-client" asp-controller="Sights" asp-action="Index" title="週邊景點">週邊景點</a></li>
+                <li><a asp-area="FP-client" asp-controller="Traffic" asp-action="Index" title="交通位置">交通位置</a></li>
+                <li><a asp-area="FP-client" asp-controller="News" asp-action="Index" title="最新消息">最新消息</a></li>
                 <li><a href="javascript: void(0)" data-toggle="modal" data-target="#myModal" title="聯絡我們">聯絡我們</a></li>
                 <li><a href="https://officialhotel.easytravel.com.tw/officialhotel_order/book_step1.aspx?HCode=H45CT2209&amp;AgCode=H45Ag68-0t9HBi" target="_blank">訂房</a></li>
 @*                <li>
@@ -48,7 +48,7 @@
                         <input type="text" name="username" id="username" tabindex="1" class="form-control" placeholder="使用者帳號" value="">
                         <input type="password" name="password" id="password" tabindex="2" class="form-control" placeholder="密碼">
                         <div class="col_2"><input type="text" class="form-control" id="inputPassword" placeholder="驗證碼"></div>
-                        <div class="col_2"><div class="mt2 left"><iframe class="" src='https://bestweb.easytravel.com.tw/sys/ck.aspx?n=20180223161838' width='60' height='26' align=bottom name='frame1' id='frame1' marginWidth=0 marginHeight=0 frameBorder=0 scrolling=no></iframe></div><span><a href="#" class="replace" title="在換一個">在換一個 <i class="fa fa-undo"></i></a></span></div>
+                        <div class="col_2"><div class="mt2 left"><iframe class="" src='https://bestweb.easytravel.com.tw/sys/ck.aspx?n=20180223161838' width='60' height='26' align=bottom name='frame1' id='frame1' marginWidth=0 marginHeight=0 frameBorder=0 scrolling=no></iframe></div><span><a href="#" class="replace" title="再換一個">再換一個 <i class="fa fa-undo"></i></a></span></div>
                         <a href="fp.html" class="forgot-password" title="忘記密碼?">忘記密碼?</a>
                         <input type="submit" name="login-submit" id="login-submit" tabindex="4" class="form-control btn btn-login" value="登入">
                         <div class="divider1 clearfix">
