@@ -1,18 +1,18 @@
-﻿<!DOCTYPE html>
+<!DOCTYPE html>
 <html lang="zh-TW">
 <head>
     <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=Edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, minimum-scale=1, user-scalable=no, target-densitydpi=medium-dpi">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="format-detection" content="telephone=no">
-    <meta name="title" content="飯店名稱">
-    <meta name="KeyWords" content="">
-    <meta name="Description" content="">
-    <meta name="copyright" content="本網頁著作權屬智凰網路科技所有！">
-    <meta name="Robots" content="INDEX,FOLLOW">
-    <meta name="revisit-after" content="1">
-    <title>飯店名稱</title>
-    <link rel="stylesheet" type="text/css" href="~/css/fp/bootstrap.min.css">
+    <meta name="description" content="@(ViewData["Description"] ?? "飯店官網")">
+    <meta name="keywords" content="@(ViewData["Keywords"] ?? "")">
+    <meta name="robots" content="index,follow">
+    <title>@(ViewData["Title"] ?? "飯店官網")</title>
+    <link rel="icon" type="image/x-icon" href="~/favicon.ico">
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link rel="stylesheet" type="text/css" href="~/css/fp/bootstrap.min.css" asp-append-version="true">
     <link rel="stylesheet" type="text/css" href="~/css/fp/font-awesome.min.css">
     <link rel="stylesheet" type="text/css" href="~/css/fp/slick.css" />
     <link rel="stylesheet" type="text/css" href="https://fonts.googleapis.com/css?family=Amiko">
@@ -25,7 +25,7 @@
     @await RenderSectionAsync("Styles", required: false)
     <!-- IE Fix for HTML5 Tags -->
     <!--[if lt IE 9]>
-    <script src="http://html5shiv.googlecode.com/svn/trunk/html5.js"></script>
+    <script src="https://oss.maxcdn.com/html5shiv/3.7.3/html5shiv.min.js"></script>
     <![endif]-->
     <!--[if lt IE 11]>
     <script>
