@@ -1,4 +1,4 @@
-﻿@{ Layout = null; }
+@{ Layout = null; }
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -19,7 +19,7 @@
                 <img src="~/img/sys/loginpic.png" alt="人物代表圖">
             </li>
             <li>
-                <p>特斯德飯店後端管理後端管理</p>
+                <p>特斯德飯店後端管理</p>
             </li>
             <li>
                 <i class="fa fa-user"></i>
