@@ -1,25 +1,22 @@
-﻿@model ErrorViewModel
+@model ErrorViewModel
 @{
-    ViewData["Title"] = "Error";
+    ViewData["Title"] = "錯誤";
+    Layout = "_Layout";
 }
-
-<h1 class="text-danger">Error.</h1>
-<h2 class="text-danger">An error occurred while processing your request.</h2>
-
-@if (Model.ShowRequestId)
-{
-    <p>
-        <strong>Request ID:</strong> <code>@Model.RequestId</code>
-    </p>
-}
-
-<h3>Development Mode</h3>
-<p>
-    Swapping to <strong>Development</strong> environment will display more detailed information about the error that occurred.
-</p>
-<p>
-    <strong>The Development environment shouldn't be enabled for deployed applications.</strong>
-    It can result in displaying sensitive information from exceptions to end users.
-    For local debugging, enable the <strong>Development</strong> environment by setting the <strong>ASPNETCORE_ENVIRONMENT</strong> environment variable to <strong>Development</strong>
-    and restarting the app.
-</p>
+<div class="container py-5">
+    <div class="text-center">
+        <h1 class="text-danger">@(Model.StatusCode ?? 500) 錯誤</h1>
+        <h2 class="text-secondary">@Model.StatusMessage</h2>
+        @if (Model.ShowRequestId)
+        {
+            <p class="mt-3"><strong>Request ID:</strong> <code>@Model.RequestId</code></p>
+        }
+        @if (Model.IsDevelopment)
+        {
+            <div class="alert alert-info mt-4 text-start">
+                <strong>開發模式</strong>：此環境會顯示詳細錯誤資訊。正式環境請將 <code>ASPNETCORE_ENVIRONMENT</code> 設為 <code>Production</code>。
+            </div>
+        }
+        <a asp-area="Sys" asp-controller="Welcome" asp-action="Index" class="btn btn-primary mt-3">返回首頁</a>
+    </div>
+</div>
