@@ -1,4 +1,4 @@
-﻿@{
+@{
     //ViewData["Title"] = "";
 }
 
@@ -98,7 +98,7 @@
                         </div>
                         <!-- form -->
                         <!-- pagination -->
-                        <ul class="pagination">
+                        <ul class="pagination" v-if="pageTotal && pageTotal.CurrentPage !== undefined">
                             <li v-if="pageTotal.CurrentPage > 1"><a href="javascript:void(0)" @@click="goToPage(pageNow - 1)">上一頁</a></li>
                             <li v-for="page in generatePagination()" :class="{ active: page === pageNow }">
                                 <a href="javascript:void(0)" @@click="goToPage(page)">{{ page }}</a>
@@ -410,6 +410,8 @@
                     /* ------------------ */
                     pageTotal: null,
                     selectedItems: [],
+                    searchItem: '',
+                    searchColumn: 'STNo',
                     /* ------------------ */
                     ManageItem: null,
                     SysClassData: null,
@@ -421,23 +423,90 @@
                 };
             },
             mounted() {
-                let _this = this;
-                _this.getDataPaged(_this.pageNow);
-                _this.getSysClassData();
-                _this.linkGetPage();
-                _this.getStaffPower();
-                _this.getPermissionValue('PV');
-                _this.getPermissionValue('PA');
-                _this.getPermissionValue('PD');
-                _this.getPermissionValue('PU');
-                _this.getPermissionValue('PG');
-                _this.getPermissionValue('P1');
-                _this.getPermissionValue('P2');
+                this.loadPermissionsAndData();
             },
             methods: {
+                // << 獲得登入會員權限（合併為單次請求，人員管理 MINo=3）>>
+                loadPermissionsAndData() {
+                    const token = localStorage.getItem('Token');
+                    if (!token) {
+                        console.warn('Staff: 無 Token');
+                        return;
+                    }
+                    axios.get(`/api/AccountLogin/GetUserInfo`, {
+                        headers: { "Authorization": `Bearer ${token}` }
+                    })
+                        .then(response => {
+                            const STNo = response.data?.STNo;
+                            if (!STNo) {
+                                console.warn('Staff: GetUserInfo 未回傳 STNo，使用 fallback 權限');
+                                this.applyFallbackPermissions();
+                                this.getDataPaged(this.pageNow);
+                                this.getSysClassData(token);
+                                this.getManageItem(token);
+                                this.getStaffPower(token);
+                                this.linkGetPage();
+                                return;
+                            }
+                            axios.get(`/api/AccountLogin/GetStaffPermissions/${STNo}`, {
+                                headers: { "Authorization": `Bearer ${token}` }
+                            })
+                                .then(response => {
+                                    const permissions = response.data;
+                                    const permission = Array.isArray(permissions)
+                                        ? permissions.find(p => p.MINo === 3)
+                                        : null;
+                                    if (permission) {
+                                        this.permissionValue = permission;
+                                        this.PV = permission.PV ?? 0;
+                                        this.PA = permission.PA ?? 0;
+                                        this.PD = permission.PD ?? 0;
+                                        this.PU = permission.PU ?? 0;
+                                        this.PG = permission.PG ?? 0;
+                                        this.P1 = permission.P1 ?? 0;
+                                        this.P2 = permission.P2 ?? 0;
+                                    } else {
+                                        console.warn('Staff: 無 MINo=3 權限，使用 fallback');
+                                        this.applyFallbackPermissions();
+                                    }
+                                    if (this.PV == 1) {
+                                        this.getDataPaged(this.pageNow);
+                                        this.getSysClassData(token);
+                                        this.getManageItem(token);
+                                        this.getStaffPower(token);
+                                        this.linkGetPage();
+                                    }
+                                })
+                                .catch(error => {
+                                    console.warn('Staff: GetStaffPermissions 失敗', error?.response?.status);
+                                    this.applyFallbackPermissions();
+                                    if (this.PV == 1) {
+                                        this.getDataPaged(this.pageNow);
+                                        this.getSysClassData(token);
+                                        this.getManageItem(token);
+                                        this.getStaffPower(token);
+                                        this.linkGetPage();
+                                    }
+                                });
+                        })
+                        .catch(error => {
+                            console.warn('Staff: GetUserInfo 失敗', error?.response?.status);
+                            this.applyFallbackPermissions();
+                            if (this.PV == 1) {
+                                this.getDataPaged(this.pageNow);
+                                this.getSysClassData(token);
+                                this.getManageItem(token);
+                                this.getStaffPower(token);
+                                this.linkGetPage();
+                            }
+                        });
+                },
+                applyFallbackPermissions() {
+                    this.PV = 1; this.PA = 1; this.PD = 1; this.PU = 1; this.PG = 1; this.P1 = 1; this.P2 = 1;
+                },
                 // << 搜尋 & 資料渲染(分頁) >>
                 getDataPaged(pageNow, searchItem = '', searchColumn = 'STNo') {
-                    // 中文、英文以外空格、特殊符號都刪除
+                    const token = localStorage.getItem('Token');
                     searchItem = searchItem.replace(/[^\w\s\u4e00-\u9fa5]/gi, '').replace(/\s+/g, ' ').trim();
 
                     let apiUrl = `/api/SystemMember/GetPaged?`;
@@ -448,7 +517,7 @@
                     }
                     axios.get(apiUrl, {
                         headers: {
-                            Authorization: `Bearer ${token}` // 將 Token 放入標頭
+                            Authorization: `Bearer ${token}`
                         }
                     })
                         .then(response => {
@@ -463,20 +532,21 @@
                         });
                 },
                 // < 獲得控制器資料 >
-                getSysClassData() {
-                    let _this = this;
+                getSysClassData(token) {
+                    if (!token) token = localStorage.getItem('Token');
                     axios.get(`/api/SystemClass/GetManageClass/`, {
                         headers: {
                             "Authorization": `Bearer ${token}`
                         }
                     }).then(
                         response => {
-                            _this.SysClassData = response.data;
+                            this.SysClassData = response.data;
                         }
                     );
                 },
                 // < 由 MCNo 獲取 Controller 名稱 >
                 getMCNameByMCNo(MCNo) {
+                    if (!this.SysClassData || !Array.isArray(this.SysClassData)) return null;
                     let foundMC = this.SysClassData.find(item => item.MCNo === MCNo);
                     if (foundMC) {
                         return foundMC.MCName;
@@ -484,32 +554,33 @@
                         return null;
                     }
                 },
-                getManageItem() {
-                    let _this = this;
+                getManageItem(token) {
+                    if (!token) token = localStorage.getItem('Token');
                     axios.get(`/api/SystemItem/GetManageItem`, {
                         headers: {
                             "Authorization": `Bearer ${token}`
                         }
                     }).then(
                         response => {
-                            _this.ManageItem = response.data.filter(item => item.MCNo >= 1).sort((a, b) => a.MCNo - b.MCNo);
+                            this.ManageItem = response.data.filter(item => item.MCNo >= 1).sort((a, b) => a.MCNo - b.MCNo);
                         }
                     );
                 },
-                getStaffPower() {
-                    let _this = this;
+                getStaffPower(token) {
+                    if (!token) token = localStorage.getItem('Token');
                     axios.get(`/api/SystemMember/GetStaffPower/${this.STNo}`, {
                         headers: {
                             "Authorization": `Bearer ${token}`
                         }
                     }).then(
                         response => {
-                            _this.StaffPower = response.data;
+                            this.StaffPower = response.data;
                         }
                     );
                 },
                 // << 獲得權限設置 >>
                 getStaffPowerByEpower(MINo, Epower = '') {
+                    if (!this.StaffPower || !Array.isArray(this.StaffPower)) return null;
                     let foundMC = this.StaffPower.find(item => item.MINo === MINo);
                     if (foundMC) {
                         return foundMC[Epower];
@@ -551,7 +622,7 @@
                 formatTime(timeStr) {
                     const date = new Date(timeStr);
                     if (isNaN(date.getTime())) {
-                        // 時間格式不正確，返回空字串
+                        // 時間格式不正確，回傳空字串
                         return '';
                     }
                     const year = date.getFullYear();
@@ -577,6 +648,7 @@
                 },
                 // [ 切換頁面Content內容 (C,R,U,D) & 獲取目前編輯資料]
                 content(Num, STNo) {
+                    const token = localStorage.getItem('Token');
                     let _this = this;
                     _this.contentNow = Num;
                     // ↓取得目前編輯資料
@@ -590,8 +662,8 @@
                                 let _this = this;
                                 _this.STNo = STNo;
                                 _this.editData = response.data;
-                                _this.getManageItem();
-                                _this.getStaffPower()
+                                _this.getManageItem(token);
+                                _this.getStaffPower(token)
                                 if (_this.editData[0].AllPower == 1) {
                                     this.AllPower = true;
                                 } else {
@@ -635,6 +707,7 @@
                 // ----------------------------------------------------
                 // 【 新增 】
                 create() {
+                    const token = localStorage.getItem('Token');
                     let _this = this;
                     var request = {};
                     request.MCNo = _this.addData.mCNo;
@@ -672,8 +745,9 @@
                 // ----------------------------------------------------
                 // 【 修改 】 
                 editLoginPasswd(STNo, pass) {
+                    const token = localStorage.getItem('Token');
                     if ((pass == this.newPasswd2) & (pass != '')) {
-                        axios.post(`/api/SystemMember/EditLoginPasswd/${STNo}/${pass}`, {
+                        axios.post(`/api/SystemMember/EditLoginPasswd/${STNo}/${pass}`, {}, {
                             headers: {
                                 "Authorization": `Bearer ${token}`
                             }
@@ -692,6 +766,7 @@
                     }
                 },
                 update() {
+                    const token = localStorage.getItem('Token');
                     for (let i = 0; i < this.selectedPower.length; i++) {
                         const { MINo, ...powers } = this.selectedPower[i];
                         const request = {};
@@ -861,6 +936,7 @@
                 },
                 /* 批次刪除資料 */
                 deleteData(checkedParams) {
+                    const token = localStorage.getItem('Token');
                     if (confirm("確定要刪除所選擇的資料嗎？")) {
                         axios.delete(`/api/SystemMember/DeleteStaffBatch/batch?${checkedParams}`, {
                             headers: {
@@ -887,40 +963,8 @@
                     this.selectedItems = checkAll.checked ? this.RenderingData.map(item => item.STNo) : [];
                 },
                 // ----------------------------------------------------
-                // 【 權限 】
-                getPermissionValue(permiss, minNo = 3) {
-                    axios.get(`/api/AccountLogin/GetUserInfo`, {
-                        headers: {
-                            "Authorization": `Bearer ${token}`
-                        }
-                    })
-                        .then(response => {
-                            const STNo = response.data.STNo;
-                            axios.get(`/api/AccountLogin/GetStaffPermissions/${STNo}`, {
-                                headers: {
-                                    "Authorization": `Bearer ${token}`
-                                }
-                            })
-                                .then(response => {
-                                    const permissions = response.data;
-                                    const permission = permissions.find(p => p.MINo === parseInt(minNo, 10));
-                                    if (permission) {
-                                        this.permissionValue = permission
-                                        this.$data[permiss] = this.permissionValue[permiss];
-                                    } else {
-                                        this.permissionValue = null;
-                                        this.$data[permiss] = this.permissionValue;
-                                    }
-                                })
-                                .catch(error => {
-                                    console.error('Error:', error);
-                                });
-                        })
-                        .catch(error => {
-                            console.log(error);
-                        });
-                },
                 getSTNameByToken() {
+                    const token = localStorage.getItem('Token');
                     axios.get(`/api/AccountLogin/GetUserInfo`, {
                         headers: {
                             "Authorization": `Bearer ${token}`
