@@ -1,6 +1,4 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using OHotel.NETCoreMVC.Controllers.API;
-using OHotel.NETCoreMVC.Controllers.FPclient;
+using Microsoft.AspNetCore.Mvc;
 using OHotel.NETCoreMVC.Models;
 using System.Diagnostics;
 
@@ -9,23 +7,10 @@ namespace OHotel.NETCoreMVC.Controllers.Sys
     [Area("Sys")]
     public class FoodController : Controller
     {
-        private readonly ILogger<HomeController> _logger;
-
-        public FoodController(ILogger<HomeController> logger)
-        {
-            _logger = logger;
-        }
-
-        public IActionResult Index()
-        {
-            return View();
-        }
-
+        public IActionResult Index() => View();
 
         [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
-        public IActionResult Error()
-        {
-            return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
-        }
+        public IActionResult Error() =>
+            View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
     }
 }
