@@ -8,12 +8,40 @@ using System.Text;
 using System.Text.Json.Serialization;
 using OHotel.NETCoreMVC.Data;
 using OHotel.NETCoreMVC.Helper;
+using OHotel.NETCoreMVC.Middleware;
 
 var builder = WebApplication.CreateBuilder(args);
 
-var connectionString = builder.Configuration.GetConnectionString("DefaultConnection");
-builder.Services.AddDbContext<ApplicationDbContext>(options =>
-    options.UseSqlServer(connectionString));
+// �q�]�w��Ū���s�u�r��A�Y���ūh�b�ҰʮɩߥX���T���~
+var connectionString = builder.Configuration.GetConnectionString("DefaultConnection")
+    ?? throw new InvalidOperationException("Connection string 'DefaultConnection' not found in configuration.");
+var dbProvider = builder.Configuration["DatabaseProvider"] ?? "SqlServer";
+
+// SQLite: �ϥε�����|�A�T�O CreateFirstAdmin �P Login �s���P�@��Ʈw
+if (string.Equals(dbProvider, "Sqlite", StringComparison.OrdinalIgnoreCase) && connectionString.Contains("ohotel.db"))
+{
+    var dbPath = Path.Combine(builder.Environment.ContentRootPath, "ohotel.db");
+    connectionString = $"Data Source={dbPath}";
+    builder.Configuration.AddInMemoryCollection(new Dictionary<string, string?>
+    {
+        ["ConnectionStrings:DefaultConnection"] = connectionString,
+        ["ConnectionStrings:SQLCD_Read_OHotel"] = connectionString
+    });
+}
+
+if (string.Equals(dbProvider, "Sqlite", StringComparison.OrdinalIgnoreCase))
+{
+    builder.Services.AddDbContext<ApplicationDbContext>(options =>
+        options.UseSqlite(connectionString));
+    builder.Services.AddScoped<IDbFunction, SqliteDbFunction>();
+}
+else
+{
+    builder.Services.AddDbContext<ApplicationDbContext>(options =>
+        options.UseSqlServer(connectionString));
+    builder.Services.AddScoped<IDbFunction, SqlServerDbFunction>();
+}
+
 builder.Services.AddDatabaseDeveloperPageExceptionFilter();
 
 builder.Services.AddDefaultIdentity<IdentityUser>(options => options.SignIn.RequireConfirmedAccount = true)
@@ -21,41 +49,54 @@ builder.Services.AddDefaultIdentity<IdentityUser>(options => options.SignIn.Requ
 
 builder.Services.AddControllersWithViews().AddJsonOptions(options =>
 {
-    //�쥻�O JsonNamingPolicy.CamelCase�A�j���Y��r��p�g�A�ڰ��n������ˡA�]��null
     options.JsonSerializerOptions.PropertyNamingPolicy = null;
-    //model josn �s���`�׶}��,�b �@��h�����p�U �n���o�U�hmodel��ƨϥήɶ}��
-    //�p�]�w options.JsonSerializerOptions.ReferenceHandler=ReferenceHandler.Preserve json �榡�|�h�X $id $value 
-    //options.JsonSerializerOptions.ReferenceHandler = null;
     options.JsonSerializerOptions.ReferenceHandler = ReferenceHandler.IgnoreCycles;
 });
 
 builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddHealthChecks();
+builder.Services.AddCors(options =>
+{
+    options.AddPolicy("CorsPolicy", policy =>
+    {
+        if (builder.Environment.IsDevelopment())
+            policy.AllowAnyOrigin().AllowAnyHeader().AllowAnyMethod();
+        else
+        {
+            var origins = builder.Configuration.GetSection("Cors:AllowedOrigins").Get<string[]>() ?? Array.Empty<string>();
+            if (origins.Length > 0)
+                policy.WithOrigins(origins).AllowAnyHeader().AllowAnyMethod();
+            else
+                policy.AllowAnyOrigin().AllowAnyHeader().AllowAnyMethod();
+        }
+    });
+});
+builder.Services.AddScoped<IVerifyHelper, VerifyHelper>();
+builder.Services.AddSingleton<IEncryptDecrypt, EncryptDecrypt>();
+builder.Services.AddSingleton<IToolsCLib, ToolsCLib>();
 
-builder.Services.AddScoped<EasyCLib.NET.Sdk.IDbFunction, SqlServerDbFunction>();//-EasyCLib.NET.Sdk ��Ʈw�s�� �̩ۨʪ`�J
-builder.Services.AddScoped<IVerifyHelper, VerifyHelper>();//-EasyCLib.NET.Sdk ��Ʈw�s�� �̩ۨʪ`�J
-builder.Services.AddSingleton<IEncryptDecrypt, EncryptDecrypt>();//�̩ۨʪ`�J EncryptDecrypt
-builder.Services.AddSingleton<IToolsCLib, ToolsCLib>();//�̩ۨʪ`�J ToolsCLib
-
-// �[�JJWT ����
+// �[�JJWT ���ҡ]�q JwtSettings Ū���A�קK�w�s�X�^
+var jwtSettings = builder.Configuration.GetSection("JwtSettings").Get<OHotel.NETCoreMVC.Models.JwtSettings>()
+    ?? throw new InvalidOperationException("JwtSettings not found in configuration.");
+var signKey = jwtSettings.SignKey;
+if (string.IsNullOrWhiteSpace(signKey) || signKey.Length < 16)
+    throw new InvalidOperationException("JwtSettings:SignKey must be at least 16 characters for security.");
 builder.Services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme).AddJwtBearer(options =>
 {
-    options.RequireHttpsMetadata = false;
-    options.TokenValidationParameters = new TokenValidationParameters()
+    options.RequireHttpsMetadata = builder.Environment.IsProduction();
+    options.TokenValidationParameters = new TokenValidationParameters
     {
-        // �z�L�o���ŧi�A�N�i�H�q "sub" ���Ȩó]�w�� User.Identity.Name
-        NameClaimType = "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier",
-        RoleClaimType = "roles",        // �z�L�o���ŧi�A�N�i�H�q "roles" ���ȡA�åi�� [Authorize] �P�_����
-        ValidateIssuer = true,        // ���ҵo���
-        ValidateAudience = false,        // �q�`���ӻݭn���� Audience
-        ValidateLifetime = true,        //���� Token �����Ĵ���
-        ValidateIssuerSigningKey = false,        // �p�G Token ���]�t key �~�ݭn���ҡA�@�볣�u��ñ���Ӥw
-        ValidIssuer = "JWT.JXINFO",        // ValidAudience = "JwtAuthDemo", // �����ҴN���ݭn��g
-        IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes("KsQvD2ROnqFOT6W4"))
+        NameClaimType = "sub",  // sub �s STNo�A�� GetUserInfo �P��� API �ϥ�
+        RoleClaimType = "roles",
+        ValidateIssuer = true,
+        ValidateAudience = false,
+        ValidateLifetime = true,
+        ValidateIssuerSigningKey = true,
+        ValidIssuer = jwtSettings.Issuer,
+        IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(signKey))
     };
 });
-//�[�J���ҪA��
 builder.Services.AddAuthorization();
-
 builder.Services.AddSwaggerGen(options =>
 {
     options.AddSecurityDefinition("BearerAuth", new OpenApiSecurityScheme
@@ -67,26 +108,36 @@ builder.Services.AddSwaggerGen(options =>
     });
     options.AddSecurityRequirement(new OpenApiSecurityRequirement
     {
-        {
-            new OpenApiSecurityScheme
-            {
-                Reference = new OpenApiReference { Type = ReferenceType.SecurityScheme, Id = "BearerAuth"}
-            },
-            new string[] { }
-        }
+        { new OpenApiSecurityScheme { Reference = new OpenApiReference { Type = ReferenceType.SecurityScheme, Id = "BearerAuth" } }, Array.Empty<string>() }
     });
 });
 
 var app = builder.Build();
 
+// Sqlite�G�Ұʮɦ۰ʪ�l�ƿ����P�ؤl�]�Y ManageClass ���š^
+if (string.Equals(dbProvider, "Sqlite", StringComparison.OrdinalIgnoreCase))
+{
+    using var scope = app.Services.CreateScope();
+    var db = scope.ServiceProvider.GetRequiredService<IDbFunction>();
+    var connStr = app.Configuration.GetConnectionString("SQLCD_Read_OHotel");
+    if (!string.IsNullOrEmpty(connStr))
+    {
+        try { SqliteDbInitializer.EnsureTablesAndSeed(db, connStr); }
+        catch { /* �����A�i���Ʈw�|���إ� */ }
+    }
+}
+
+// ����ҥ~�B�z�]���b�̫e�^
+app.UseMiddleware<ExceptionHandlingMiddleware>();
+
 if (app.Environment.IsDevelopment())
 {
     app.UseMigrationsEndPoint();
-    app.UseSwagger(); // ���F��K�ާ@API�[�J
-    app.UseSwaggerUI(options => // ���F��K�ާ@API�ӥ[�J�ϥΪ�UI����
+    app.UseSwagger();
+    app.UseSwaggerUI(options =>
     {
         options.SwaggerEndpoint("/swagger/v1/swagger.json", "API V1");
-        options.RoutePrefix = "api/swagger"; /*�o��]�w�s��*/
+        options.RoutePrefix = "api/swagger";
     });
 }
 else
@@ -95,42 +146,69 @@ else
     app.UseHsts();
 }
 
-app.UseStaticFiles(); //���\�ɮ�Ū��
-
-// If there are calls to app.UseRouting() and app.UseEndpoints(...), the call to app.UseCors() must go between them.
-app.UseRouting();
+app.UseStatusCodePagesWithReExecute("/Home/Error", "?statusCode={0}");
 app.UseHttpsRedirection();
-app.UseCors(options =>
+app.UseStaticFiles();
+app.UseRouting();
+app.Use(async (context, next) =>
 {
-    options.AllowAnyOrigin()
-        .AllowAnyHeader()
-        .AllowAnyMethod();
-}); //�ϥθ���� // ���FAPI�[�J
+    context.Response.Headers.Append("X-Content-Type-Options", "nosniff");
+    context.Response.Headers.Append("X-Frame-Options", "SAMEORIGIN");
+    context.Response.Headers.Append("Referrer-Policy", "strict-origin-when-cross-origin");
+    await next();
+});
+app.UseCors("CorsPolicy");
 
-app.UseAuthentication(); //����
-app.UseAuthorization(); //���v
+app.UseAuthentication();
+app.UseAuthorization();
 
-//app.MapAreaControllerRoute( // �ζH�x��  (�¼g�k �|�d FP-client)
-//    name: "client_route",
-//    areaName: "FP-client",
-//    pattern: "{area:exists}/{controller=Home}/{action=Index}/{id?}",
-//    defaults: new { action = "Index" }
-//);
+app.MapControllers(); // API �ݩʸ��Ѷ��u����ǲθ���
 
-app.MapControllerRoute( // �ζH�x��
+// Health Check ���I�]�ѭt�����ž��κʱ��ϥΡ^
+app.MapHealthChecks("/health");
+
+// �ƴ��G�̤p API ���I�]�Y Controller ���� 404 �ɥi�Ρ^
+app.MapPost("/api/account-login/create-first-admin", async (IDbFunction db, IConfiguration config) =>
+{
+    var connStr = config.GetConnectionString("SQLCD_Read_OHotel");
+    if (string.IsNullOrEmpty(connStr)) return Results.BadRequest("Database not configured");
+    var dbProvider = config["DatabaseProvider"] ?? "";
+    try
+    {
+        db.DbConnect(connStr);
+        if (string.Equals(dbProvider, "Sqlite", StringComparison.OrdinalIgnoreCase))
+        {
+            var createSql = @"CREATE TABLE IF NOT EXISTS Staff (STNo INTEGER PRIMARY KEY AUTOINCREMENT, STName TEXT, LoginName TEXT, LoginPasswd TEXT, Tel TEXT, EMail TEXT, State INTEGER DEFAULT 0, AllPower INTEGER DEFAULT 0, LoginTime TEXT, LAmount INTEGER DEFAULT 0, CTime TEXT, MTime TEXT, MSNo INTEGER DEFAULT 0)";
+            if (!db.AlterDb(createSql)) { db.DbClose(); return Results.StatusCode(500); }
+            db.DbClose();
+            OHotel.NETCoreMVC.Data.SqliteDbInitializer.EnsureTablesAndSeed(db, connStr);
+            db.DbConnect(connStr);
+        }
+        if (!db.SelectDbDataView("SELECT COUNT(*) as cnt FROM Staff", "Staff")) { db.DbClose(); return Results.StatusCode(500); }
+        if (db.SqlDataView.Count > 0 && Convert.ToInt32(db.SqlDataView[0]["cnt"]) > 0) { db.DbClose(); return Results.BadRequest("�w���޲z��"); }
+        var enc = OHotelCLib.Alenher.EandD_Reply.StaffSet("admin123", 1, 1).Replace("'", "''");
+        var now = DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss");
+        var ins = $"INSERT INTO Staff (STName, LoginName, LoginPasswd, Tel, EMail, State, AllPower, LoginTime, LAmount, CTime, MTime, MSNo) VALUES ('�t�κ޲z��', 'admin', '{enc}', '', '', 0, 1, '{now}', 0, '{now}', '{now}', 0)";
+        if (!db.AlterDb(ins)) { db.DbClose(); return Results.StatusCode(500); }
+        db.DbClose();
+        return Results.Ok(new { success = true, loginName = "admin", password = "admin123", hint = "�Ц� /Sys/Login �n�J" });
+    }
+    catch { db.DbClose(); return Results.StatusCode(500); }
+});
+app.MapControllerRoute(
     name: "client_route",
     pattern: "{controller=Home}/{action=Index}/{id?}",
     defaults: new { area = "FP-client" }
 );
 
-app.MapAreaControllerRoute( // �޲z�̤���
+app.MapAreaControllerRoute(
     name: "dev_route",
     areaName: "Sys",
     pattern: "Sys/{controller=Welcome}/{action=Index}/{id?}",
     defaults: new { action = "Index" }
 );
 
-app.MapControllerRoute( // �q�{
+app.MapControllerRoute(
     name: "default",
     pattern: "{controller=Home}/{action=Index}/{id?}");
 
