@@ -1,27 +1,27 @@
 /*
- * fixedScroll.js 滚动固定插件
+ * fixedScroll.js 捲動固定插件
  * @DH   
  * https://denghao.me
  * 
- * 示例：
+ * 範例：
  * $('.box').fixedScroll()
  */
 
 ;(function () {
     let fixedScroll = function ($fixedEl, opts) {
         this.defaults = {
-            navEls: '', //nav (注意: navEls和hookEls两个参数必须成对出现)
-            hookEls: '', //nav要滚动到的对应元素
-            hookOffset: 0, //hook区域顶部偏移量
-            offset: 0, //固定元素顶部偏移量
-            stickEndEl: '', //固定结束位置的元素
+            navEls: '', //nav (注意: navEls 和 hookEls 兩個參數必須成對出現)
+            hookEls: '', //nav 要捲動到的對應元素
+            hookOffset: 0, //hook 區域頂部偏移量
+            offset: 0, //固定元素頂部偏移量
+            stickEndEl: '', //固定結束位置的元素
             callback: ''
         };
         $.extend(this, this.defaults, opts);
         this.flag = true;
         this.stickTop = 0; //固定元素的原始位置
-        this.init_stickTop = 0; //用于重计算高度
-        this.stickBottom = 9999999; //固定元素的结束位置
+        this.init_stickTop = 0; //用於重新計算高度
+        this.stickBottom = 9999999; //固定元素的結束位置
         this.fixedEl = $fixedEl; //固定元素
         this.fixedElH = $fixedEl.height();
         this.fixedElW = $fixedEl.width();
@@ -41,7 +41,7 @@
             if (this.stickEndEl.length > 0) {
                 this.stickBottom = this.stickEndEl.offset().top;
             }
-            // 限定起始位的距顶高度
+            // 限定起始位的距頂高度
             this.distance = this.stickBottom - this.stickTop - this.fixedElH - this.offset / 2;
             this.calcArea();
             this.flag && this.events();
@@ -78,7 +78,7 @@
             }
         },
 
-        // 动态计算高度
+        // 動態計算高度
         resizeHeight: function (hasNewTop) {
             if (this.fixedEl.length > 0) {
                 this.stickTop = hasNewTop ? this.fixedEl.offset().top : this.init_stickTop;
@@ -90,7 +90,7 @@
             this.calcArea();
         },
 
-        // 计算滚动区
+        // 計算捲動區
         calcArea: function () {
             if (this.hookEls.length <= 0) return;
             let areas = [];
@@ -102,7 +102,7 @@
             this.hookArea = areas;
         },
 
-        // 区域判断
+        // 區域判斷
         hookScroll: function () {
             var t = this.winEl.scrollTop();
             for (var i in this.hookArea) {
@@ -114,7 +114,7 @@
             }
         },
 
-        // nav状态
+        // nav 狀態
         navStatus: function (i) {
             if (i || +i === 0) {
                 this.navEls.eq(i).addClass('active').siblings().removeClass('active');
@@ -123,7 +123,7 @@
             }
         },
 
-        // 滚动到指定位置
+        // 捲動到指定位置
         refresh: function (i) {
             let top = this.hookArea[i][0] - this.hookOffset;
             this.calcArea();
@@ -150,7 +150,7 @@
 
         events: function () {
             let _this = this;
-            // 切换nav
+            // 切換 nav
             if (_this.navEls.length > 0) {
                 this.navEls.on('click', function () {
                     let i = $(this).index();
@@ -162,7 +162,7 @@
                     }, 300);
                 })
             }
-            // 滚动监听
+            // 捲動監聽
             this.winEl.on("scroll", function () {
                 (_this.fixedEl.length > 0) && _this.stickHandle();
                 (_this.hookEls.length > 0 && !_this.isClickSwitch) && _this.hookScroll();
@@ -176,7 +176,7 @@
         return drag;
     }
 
-    //兼容模块
+    // 相容模組
     if (typeof module !== 'undefined' && typeof exports === 'object') {
         module.exports = fixedScroll;
     } else if (typeof define === 'function' && (define.amd || define.cmd)) {
